Factor out layer vis and its state
