+++ src/components/map/DataMap.js
@@ -1,22 +1,21 @@
 // DataMap: Component that displays a map with data.
 
-import React, { useMemo, useState } from "react";
+import React, { useMemo } from "react";
 import PropTypes from "prop-types";
-import { LayersControl } from "react-leaflet";
 import _ from "lodash";
 import flow from "lodash/fp/flow";
 import map from "lodash/fp/map";
 import compact from "lodash/fp/compact";
 
 import { BCBaseMap } from "pcic-react-leaflet-components";
-import "@/components/map/DataMap.css";
 import { useConfigContext } from "@/state/context-hooks/use-config-context";
 import MapSpinner from "@/components/map/MapSpinner";
 import useBaseline from "@/state/query-hooks/use-baseline";
 import useMonthly from "@/state/query-hooks/use-monthly";
 import { formatDate } from "@/components/utils";
-import StationDataMarkersPane from "@/components/map/StationDataMarkersPane";
-import StationLocationMarkersPane from "@/components/map/StationLocationMarkersPane";
+import MarkerLayers from "@/components/map/MarkerLayers";
+
+import "@/components/map/DataMap.css";
 
 export default function DataMap({ dataset, variable, date }) {
   // Server state
@@ -31,19 +30,6 @@ export default function DataMap({ dataset, variable, date }) {
     isPending: monthlyIsPending,
     isError: monthlyIsError,
   } = useMonthly(variable, date);
-
-  // App state
-  const [visibleLayers, setVisibleLayers] = useState({
-    stationDataValueMarkers: true,
-    monthlyStationLocationMarkers: false,
-    baselineStationLocationMarkers: false,
-  });
-  const getLayerVisibility = (layer) => visibleLayers[layer];
-  const setLayerVisibility = (layer, visibility) =>
-    setVisibleLayers({
-      ...visibleLayers,
-      [layer]: visibility,
-    });
 
   const isBaseline = dataset === "baseline";
   const isMonthly = dataset === "monthly";
@@ -82,52 +68,6 @@ export default function DataMap({ dataset, variable, date }) {
     )(baseline);
   }, [dataset, monthly, baseline]);
 
-  // TODO: Consider useMemo for each item
-  // Unordered marker layers
-  // Each overlay group is enclosed in a Pane, which controls the ordering
-  // of the layers on the map. Rendering them in a particular order (see below)
-  // within the map controls which layer overlays which.
-  const markerLayersById = {
-    data: (
-      <StationDataMarkersPane
-        key={"data"}
-        layerName={"stationDataValueMarkers"}
-        variable={variable}
-        dataset={dataset}
-        date={date}
-        stations={stationsForDataset}
-        getVisibility={getLayerVisibility}
-        setVisibility={setLayerVisibility}
-      />
-    ),
-
-    monthly: (
-      <StationLocationMarkersPane
-        key={"monthly"}
-        layerName={"monthlyStationLocationMarkers"}
-        variable={variable}
-        dataset={"monthly"}
-        date={date}
-        stations={monthly}
-        getVisibility={getLayerVisibility}
-        setVisibility={setLayerVisibility}
-      />
-    ),
-
-    baseline: (
-      <StationLocationMarkersPane
-        key={"baseline"}
-        layerName={"baselineStationLocationMarkers"}
-        variable={variable}
-        dataset={"baseline"}
-        date={date}
-        stations={baseline}
-        getVisibility={getLayerVisibility}
-        setVisibility={setLayerVisibility}
-      />
-    ),
-  };
-
   const makeContent = () => {
     if (baselineIsPending || monthlyIsPending) {
       return <MapSpinner>Loading data...</MapSpinner>;
@@ -147,15 +87,15 @@ export default function DataMap({ dataset, variable, date }) {
       );
     }
 
-    // Render marker layers in order defined by config.
-    // Because each layer group is enclosed in a Pane, the ordering controls
-    // which layer overlays which. Later-rendered Panes overlay earlier ones.
-    // The ordering also controls the order the layer is listed
-    // in the LayersControl.
     return (
-      <LayersControl position="topright">
-        {config.map.markerLayers.order.map((id) => markerLayersById[id])}
-      </LayersControl>
+      <MarkerLayers
+        dataset={dataset}
+        variable={variable}
+        date={date}
+        baseline={baseline}
+        monthly={monthly}
+        stationsForDataset={stationsForDataset}
+      />
     );
   };
